mv writeEpilogue method to BondsetData
