+++ src/module/item/item.ts
@@ -1,9 +1,6 @@
 import type { ConfiguredData } from '@league-of-foundry-developers/foundry-vtt-types/src/types/helperTypes'
 import type { DocumentSubTypes } from '../../types/helperTypes'
-import { getFoundryMoveByDfId } from '../dataforged'
-import { IronswornPrerollDialog } from '../rolls'
 import type { ItemDataProperties } from './config'
-import { BondsetDataPropertiesData } from './subtypes/bondset'
 
 /**
  * Extend the base Item entity
@@ -34,32 +31,6 @@ export class IronswornItem<
 		subtype: T
 	): this is IronswornItem<T> {
 		return IronswornItem.assert(this, subtype)
-	}
-
-	/**
-	 * Bondset methods
-	 */
-
-	async writeEpilogue() {
-		if (this.type !== 'bondset') return
-		const system = this.system as BondsetDataPropertiesData
-
-		const move = await getFoundryMoveByDfId(
-			'Ironsworn/Moves/Relationship/Write_Your_Epilogue'
-		)
-		if (move == null) throw new Error('Problem loading write-epilogue move')
-
-		const progress = Math.floor(Object.values(system.bonds).length / 4)
-		void IronswornPrerollDialog.showForOfficialMove(
-			'Ironsworn/Moves/Relationship/Write_Your_Epilogue',
-			{
-				actor: this.actor ?? undefined,
-				progress: {
-					source: game.i18n.localize('IRONSWORN.ITEMS.TypeBond'),
-					value: progress
-				}
-			}
-		)
 	}
 }
