+++ src/module/item/subtypes/bondset.ts
@@ -1,4 +1,6 @@
+import { getFoundryMoveByDfId } from '../../dataforged'
 import type { DataSchema } from '../../fields/utils'
+import { IronswornPrerollDialog } from '../../rolls'
 import type { IronswornItem } from '../item'
 
 export class BondsetData extends foundry.abstract.DataModel<
@@ -6,6 +8,25 @@ export class BondsetData extends foundry.abstract.DataModel<
 	IronswornItem<'bondset'>
 > {
 	static _enableV10Validation = true
+
+	async writeEpilogue() {
+		const move = await getFoundryMoveByDfId(
+			'Ironsworn/Moves/Relationship/Write_Your_Epilogue'
+		)
+		if (move == null) throw new Error('Problem loading write-epilogue move')
+
+		const progress = Math.floor(this.bonds.length / 4)
+		void IronswornPrerollDialog.showForOfficialMove(
+			'Ironsworn/Moves/Relationship/Write_Your_Epilogue',
+			{
+				actor: this.parent.actor ?? undefined,
+				progress: {
+					source: game.i18n.localize('IRONSWORN.ITEMS.TypeBond'),
+					value: progress
+				}
+			}
+		)
+	}
 
 	static override defineSchema(): DataSchema<BondsetDataSourceData> {
 		const fields = foundry.data.fields
@@ -19,7 +40,7 @@ export class BondsetData extends foundry.abstract.DataModel<
 		}
 	}
 }
-export interface BondsetData extends BondsetDataPropertiesData {}
+export interface BondsetData extends BondsetDataSourceData {}
 
 export interface Bond {
 	name: string
@@ -29,8 +50,6 @@ export interface Bond {
 export interface BondsetDataSourceData {
 	bonds: Bond[]
 }
-export interface BondsetDataPropertiesData extends BondsetDataSourceData {}
-
 export interface BondsetDataSource {
 	type: 'bondset'
 	data: BondsetDataSourceData
@@ -38,6 +57,6 @@ export interface BondsetDataSource {
 }
 export interface BondsetDataProperties {
 	type: 'bondset'
-	data: BondsetDataPropertiesData
-	system: BondsetDataPropertiesData
+	data: BondsetData
+	system: BondsetData
 }
