+++ src/module/vue/components/bonds.vue
@@ -2,11 +2,19 @@
 	<div class="flexcol nogrow">
 		<div class="flexrow">
 			<h4>{{ $t('IRONSWORN.ITEMS.TypeBond') }}</h4>
-			<IronBtn block nogrow icon="fa:pen-to-square" @click="editBonds" />
-			<IronBtn block nogrow icon="ironsworn:d10-tilt" @click="rollBonds" />
+			<IronBtn
+				block
+				nogrow
+				icon="fa:pen-to-square"
+				@click="$bonds.sheet?.render(true)" />
+			<IronBtn
+				block
+				nogrow
+				icon="ironsworn:d10-tilt"
+				@click="$bonds.system.writeEpilogue()" />
 		</div>
 		<ProgressTrack
-			:ticks="bondcount"
+			:ticks="bonds.system.bonds.length"
 			:rank="null"
 			:compact-progress="props.compactProgress" />
 	</div>
@@ -14,6 +22,7 @@
 
 <script setup lang="ts">
 import { inject, computed } from 'vue'
+import type { IronswornItem } from '../../item/item'
 import { $ActorKey, ActorKey } from '../provisions'
 import IronBtn from './buttons/iron-btn.vue'
 import ProgressTrack from './progress/progress-track.vue'
@@ -23,24 +32,15 @@ const props = defineProps<{ compactProgress?: boolean }>()
 const actor = inject(ActorKey)
 const $actor = inject($ActorKey)
 
-const bonds = computed(() => {
-	return actor?.value?.items.find(
-		(x) => x.type === 'bondset'
-	) as ItemSource<'bondset'>
-})
-const bondcount = computed(() => {
-	if (bonds?.value?.system.bonds == null) return 0
-	return Object.values(bonds?.value?.system.bonds).length
-})
-
-function editBonds() {
-	const item = $actor?.items.get(bonds?.value?._id as string)
-	item?.sheet?.render(true)
-}
-function rollBonds() {
-	const item = $actor?.items.get(bonds?.value?._id as string)
-	item?.writeEpilogue()
-}
+const bonds = computed(
+	() =>
+		actor?.value?.items.find(
+			(x) => x.type === 'bondset'
+		) as ItemSource<'bondset'>
+)
+const $bonds = computed(
+	() => $actor?.itemTypes.bondset[0] as IronswornItem<'bondset'>
+)
 </script>
 
 <style lang="scss" scoped>
